third_party/blink/web_tests/external/wpt/html/semantics/document-metadata/the-meta-element/color-scheme/meta-color-scheme-attribute-changes-expected.txt

This is a testharness.js-based test.
FAIL Meta color-scheme initially 'dark'. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
PASS Removed name attribute from meta color-scheme.
FAIL Set meta name to color-scheme. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
PASS Set content attribute of meta color-scheme to empty string.
PASS Set content attribute of meta color-scheme to an invalid value.
FAIL Set content attribute of meta color-scheme to 'light'. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "light"
FAIL Set content attribute of meta color-scheme to 'dark'. assert_equals: Root element's color-scheme should be 'normal' expected "normal" but got "dark"
PASS Removed the content attribute of meta color-scheme.
Harness: the test ran to completion.
